Support ECDH with OpenSSL-1.0.2
This  patch requires the matching patch in libp11.

This is a hack to allow one to build libp11 and opensc's engine_pkcs11 using
OpenSSL-1.0.2. But OpenSSL-1.0.2 does not allow access to the ECDH_METHOD structure,
and doesn not provide access functions to change the compute_key function.
(OpenSSL-1.1 does provided the needed functions).

So this hack is only for use with 1.0.2 and only for those who are willing to debug
the code if OpenSSL changes anything that is hidden.

You must define BUILD_ECDH_102.

CPPFLAGS="-DBUILD_ECDH_102"
export CPPFLAGS
configure ...